pyproject.toml
[project]
name = "imnotmusic"
version = "0.1.0"
description = "A self hosted music database with custom ranking systems"
authors = [
    {name = "Nordanir"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2 (>=2.9.10,<3.0.0)",
    "wand (>=0.6.13,<0.7.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
